+++ frontend/src/components/CategoryList/CategoryList.tsx
@@ -12,7 +12,8 @@ import { userState } from '@/recoil/atoms';
 import SubscribedCategoryItem from '@/components/SubscribedCategoryItem/SubscribedCategoryItem';
 import UnsubscribedCategoryItem from '@/components/UnsubscribedCategoryItem/UnsubscribedCategoryItem';
 
-import { API, CACHE_KEY } from '@/constants';
+import { CACHE_KEY } from '@/constants';
+import { API } from '@/constants/api';
 
 import categoryApi from '@/api/category';
 import subscriptionApi from '@/api/subscription';
