+++ frontend/src/constants/api.ts
@@ -1,0 +1,7 @@
+const API = {
+  CATEGORY_GET_SIZE: 4,
+};
+
+const API_URL = process.env.API_URL;
+
+export { API, API_URL };
