+++ frontend/src/constants/index.ts
@@ -1,9 +1,3 @@
-const API = {
-  CATEGORY_GET_SIZE: 4,
-};
-
-const API_URL = process.env.API_URL;
-
 const ATOM_KEY = {
   SNACK_BAR: 'snackBarState',
   SIDE_BAR: 'sideBarState',
@@ -47,14 +41,4 @@ const VALIDATION_SIZE = {
   CATEGORY_NAME_MAX_LENGTH: 20,
 };
 
-export {
-  API,
-  API_URL,
-  ATOM_KEY,
-  CACHE_KEY,
-  CALENDAR,
-  SELECTOR_KEY,
-  STORAGE_KEY,
-  PATH,
-  VALIDATION_SIZE,
-};
+export { ATOM_KEY, CACHE_KEY, CALENDAR, SELECTOR_KEY, STORAGE_KEY, PATH, VALIDATION_SIZE };
